Enviando entidades ao repo

+++ src/main/resources/db/migration/V1__create-table-Cliente.sql
@@ -1,10 +1,9 @@
  create table cliente_tb(
 
-    id bigint not null auto_increment,
+    id int not null auto_increment,
     nome varchar(100) not null,
     email varchar(100) not null unique,
     
-
     primary key(id)
 
 );
